Remove boost::visitor from RSModel.
