+++ src/mlpack/methods/range_search/CMakeLists.txt
@@ -8,6 +8,7 @@ set(SOURCES
   range_search_stat.hpp
   rs_model.hpp
   rs_model_impl.hpp
+  rs_model.cpp
 )
 
 # Add directory name to sources.
