+++ src/mlpack/methods/range_search/rs_model_impl.hpp
@@ -20,295 +20,87 @@
 namespace mlpack {
 namespace range {
 
-/**
- * Initialize the RSModel with the given tree type and whether or not a random
- * basis should be used.
- */
-RSModel::RSModel(TreeTypes treeType, bool randomBasis) :
-    treeType(treeType),
-    leafSize(0),
-    randomBasis(randomBasis),
-    rSearch(NULL)
-{
-  // Nothing to do.
-}
-
-// Copy constructor.
-RSModel::RSModel(const RSModel& other) :
-    treeType(other.treeType),
-    leafSize(other.leafSize),
-    randomBasis(other.randomBasis),
-    q(other.q),
-    rSearch(other.rSearch->Clone())
-{
-  // Nothing to do.
-}
-
-// Move constructor.
-RSModel::RSModel(RSModel&& other) :
-    treeType(other.treeType),
-    leafSize(other.leafSize),
-    randomBasis(other.randomBasis),
-    q(std::move(other.q)),
-    rSearch(std::move(other.rSearch))
-{
-  // Reset other model.
-  other.treeType = TreeTypes::KD_TREE;
-  other.leafSize = 0;
-  other.randomBasis = false;
-}
-
-// Copy operator.
-RSModel& RSModel::operator=(const RSModel& other)
-{
-  if (this != &other)
-  {
-    delete rSearch;
-
-    treeType = other.treeType;
-    leafSize = other.leafSize;
-    randomBasis = other.randomBasis;
-    q = other.q;
-    rSearch = other.rSearch->Clone();
-  }
-
-  return *this;
-}
-
-// Move operator.
-RSModel& RSModel::operator=(RSModel&& other)
-{
-  if (this != &other)
-  {
-    delete rSearch;
-
-    treeType = other.treeType;
-    leafSize = other.leafSize;
-    randomBasis = other.randomBasis;
-    q = std::move(other.q);
-    rSearch = std::move(other.rSearch);
-
-    other.treeType = TreeTypes::KD_TREE;
-    other.leafSize = 0;
-    other.randomBasis = false;
-  }
-
-  return *this;
-}
-
-// Clean memory, if necessary.
-RSModel::~RSModel()
-{
-  delete rSearch;
-}
-
-void RSModel::BuildModel(arma::mat&& referenceSet,
-                         const size_t leafSize,
-                         const bool naive,
-                         const bool singleMode)
-{
-  // Initialize random basis if necessary.
-  if (randomBasis)
-  {
-    Log::Info << "Creating random basis..." << std::endl;
-    math::RandomBasis(q, referenceSet.n_rows);
-  }
-
-  this->leafSize = leafSize;
-
-  // Clean memory, if necessary.
-  delete rSearch;
-
-  // Do we need to modify the reference set?
-  if (randomBasis)
-    referenceSet = q * referenceSet;
-
-  if (!naive)
-  {
-    Timer::Start("tree_building");
-    Log::Info << "Building reference tree..." << std::endl;
-  }
-
-  switch (treeType)
-  {
-    case KD_TREE:
-      rSearch = new LeafSizeRSWrapper<tree::KDTree>(naive, singleMode);
-      break;
-
-    case COVER_TREE:
-      rSearch = new RSWrapper<tree::StandardCoverTree>(naive, singleMode);
-      break;
-
-    case R_TREE:
-      rSearch = new RSWrapper<tree::RTree>(naive, singleMode);
-      break;
-
-    case R_STAR_TREE:
-      rSearch = new RSWrapper<tree::RStarTree>(naive, singleMode);
-      break;
-
-    case BALL_TREE:
-      rSearch = new LeafSizeRSWrapper<tree::BallTree>(naive, singleMode);
-      break;
-
-    case X_TREE:
-      rSearch = new RSWrapper<tree::XTree>(naive, singleMode);
-      break;
-
-    case HILBERT_R_TREE:
-      rSearch = new RSWrapper<tree::HilbertRTree>(naive, singleMode);
-      break;
-
-    case R_PLUS_TREE:
-      rSearch = new RSWrapper<tree::RPlusTree>(naive, singleMode);
-      break;
-
-    case R_PLUS_PLUS_TREE:
-      rSearch = new RSWrapper<tree::RPlusPlusTree>(naive, singleMode);
-      break;
-
-    case VP_TREE:
-      rSearch = new RSWrapper<tree::VPTree>(naive, singleMode);
-      break;
-
-    case RP_TREE:
-      rSearch = new RSWrapper<tree::RPTree>(naive, singleMode);
-      break;
-
-    case MAX_RP_TREE:
-      rSearch = new RSWrapper<tree::MaxRPTree>(naive, singleMode);
-      break;
-
-    case UB_TREE:
-      rSearch = new RSWrapper<tree::UBTree>(naive, singleMode);
-      break;
-
-    case OCTREE:
-      rSearch = new LeafSizeRSWrapper<tree::Octree>(naive, singleMode);
-      break;
-  }
-
-  rSearch->Train(std::move(referenceSet), leafSize);
-
-  if (!naive)
-  {
-    Timer::Stop("tree_building");
-    Log::Info << "Tree built." << std::endl;
-  }
-}
-
-// Perform range search.
-void RSModel::Search(arma::mat&& querySet,
-                     const math::Range& range,
-                     std::vector<std::vector<size_t>>& neighbors,
-                     std::vector<std::vector<double>>& distances)
-{
-  // We may need to map the query set randomly.
-  if (randomBasis)
-    querySet = q * querySet;
-
-  Log::Info << "Search for points in the range [" << range.Lo() << ", "
-      << range.Hi() << "] with ";
-  if (!Naive() && !SingleMode())
-    Log::Info << "dual-tree " << TreeName() << " search..." << std::endl;
-  else if (!Naive())
-    Log::Info << "single-tree " << TreeName() << " search..." << std::endl;
-  else
-    Log::Info << "brute-force (naive) search..." << std::endl;
-
-  rSearch->Search(std::move(querySet), range, neighbors, distances, leafSize);
-}
-
-// Perform range search (monochromatic case).
-void RSModel::Search(const math::Range& range,
-                     std::vector<std::vector<size_t>>& neighbors,
-                     std::vector<std::vector<double>>& distances)
-{
-  Log::Info << "Search for points in the range [" << range.Lo() << ", "
-      << range.Hi() << "] with ";
-  if (!Naive() && !SingleMode())
-    Log::Info << "dual-tree " << TreeName() << " search..." << std::endl;
-  else if (!Naive())
-    Log::Info << "single-tree " << TreeName() << " search..." << std::endl;
-  else
-    Log::Info << "brute-force (naive) search..." << std::endl;
-
-  rSearch->Search(range, neighbors, distances);
-}
-
-// Get the name of the tree type.
-std::string RSModel::TreeName() const
-{
-  switch (treeType)
-  {
-    case KD_TREE:
-      return "kd-tree";
-    case COVER_TREE:
-      return "cover tree";
-    case R_TREE:
-      return "R tree";
-    case R_STAR_TREE:
-      return "R* tree";
-    case BALL_TREE:
-      return "ball tree";
-    case X_TREE:
-      return "X tree";
-    case HILBERT_R_TREE:
-      return "Hilbert R tree";
-    case R_PLUS_TREE:
-      return "R+ tree";
-    case R_PLUS_PLUS_TREE:
-      return "R++ tree";
-    case VP_TREE:
-      return "vantage point tree";
-    case RP_TREE:
-      return "random projection tree (mean split)";
-    case MAX_RP_TREE:
-      return "random projection tree (max split)";
-    case UB_TREE:
-      return "UB tree";
-    case OCTREE:
-      return "octree";
-    default:
-      return "unknown tree";
-  }
-}
-
-// Clean memory.
-void RSModel::CleanMemory()
-{
-  delete rSearch;
-}
-
-//! Default Bichromatic range search on the given RSType instance.
 template<template<typename TreeMetricType,
                   typename TreeStatType,
                   typename TreeMatType> class TreeType>
-void BiSearchVisitor::operator()(RSTypeT<TreeType>* rs) const
+void RSWrapper<TreeType>::Train(arma::mat&& referenceSet,
+                                const size_t /* leafSize */)
 {
-  if (rs)
-    return rs->Search(querySet, range, neighbors, distances);
-  throw std::runtime_error("no range search model initialized");
+  rs.Train(std::move(referenceSet));
 }
 
-//! Bichromatic range search on the given RSType considering the leafSize.
-template<typename RSType>
-void BiSearchVisitor::SearchLeaf(RSType* rs) const
+template<template<typename TreeMetricType,
+                  typename TreeStatType,
+                  typename TreeMatType> class TreeType>
+void RSWrapper<TreeType>::Search(arma::mat&& querySet,
+                                 const math::Range& range,
+                                 std::vector<std::vector<size_t>>& neighbors,
+                                 std::vector<std::vector<double>>& distances,
+                                 const size_t /* leafSize */)
 {
-  if (!rs->Naive() && !rs->SingleMode())
+  rs.Search(std::move(querySet), range, neighbors, distances);
+}
+
+template<template<typename TreeMetricType,
+                  typename TreeStatType,
+                  typename TreeMatType> class TreeType>
+void RSWrapper<TreeType>::Search(const math::Range& range,
+                                 std::vector<std::vector<size_t>>& neighbors,
+                                 std::vector<std::vector<double>>& distances)
+{
+  rs.Search(range, neighbors, distances);
+}
+
+template<template<typename TreeMetricType,
+                  typename TreeStatType,
+                  typename TreeMatType> class TreeType>
+void LeafSizeRSWrapper<TreeType>::Train(arma::mat&& referenceSet,
+                                        const size_t leafSize)
+{
+  if (rs.Naive())
+  {
+    rs.Train(std::move(referenceSet));
+  }
+  else
+  {
+    std::vector<size_t> oldFromNewReferences;
+    typename decltype(rs)::Tree* tree =
+        new typename decltype(rs)::Tree(std::move(referenceSet),
+                                        oldFromNewReferences,
+                                        leafSize);
+    rs.Train(tree);
+
+    // Give the model ownership of the tree and the mappings.
+    rs.treeOwner = true;
+    rs.oldFromNewReferences = std::move(oldFromNewReferences);
+  }
+}
+
+template<template<typename TreeMetricType,
+                  typename TreeStatType,
+                  typename TreeMatType> class TreeType>
+void LeafSizeRSWrapper<TreeType>::Search(
+    arma::mat&& querySet,
+    const math::Range& range,
+    std::vector<std::vector<size_t>>& neighbors,
+    std::vector<std::vector<double>>& distances,
+    const size_t leafSize)
+{
+  if (!rs.Naive() && !rs.SingleMode())
   {
     // Build a second tree and search.
     Timer::Start("tree_building");
     Log::Info << "Building query tree..." << std::endl;
     std::vector<size_t> oldFromNewQueries;
-    typename RSType::Tree queryTree(std::move(querySet), oldFromNewQueries,
-        leafSize);
+    typename decltype(rs)::Tree queryTree(std::move(querySet),
+                                          oldFromNewQueries,
+                                          leafSize);
     Log::Info << "Tree built." << std::endl;
     Timer::Stop("tree_building");
 
     std::vector<std::vector<size_t>> neighborsOut;
     std::vector<std::vector<double>> distancesOut;
-    rs->Search(&queryTree, range, neighborsOut, distancesOut);
+    rs.Search(&queryTree, range, neighborsOut, distancesOut);
 
     // Remap the query points.
     neighbors.resize(queryTree.Dataset().n_cols);
@@ -320,44 +112,8 @@ void BiSearchVisitor::SearchLeaf(RSType* rs) const
     }
   }
   else
-    rs->Search(querySet, range, neighbors, distances);
-}
-
-//! Save parameters for Train.
-inline TrainVisitor::TrainVisitor(arma::mat&& referenceSet,
-                                  const size_t leafSize) :
-    referenceSet(std::move(referenceSet)),
-    leafSize(leafSize)
-{}
-
-//! Default Train on the given RSType instance.
-template<template<typename TreeMetricType,
-                  typename TreeStatType,
-                  typename TreeMatType> class TreeType>
-void TrainVisitor::operator()(RSTypeT<TreeType>* rs) const
-{
-  if (rs)
-    return rs->Train(std::move(referenceSet));
-  throw std::runtime_error("no range search model initialized");
-}
-
-//! Train on the given RSType considering the leafSize.
-template<typename RSType>
-void TrainVisitor::TrainLeaf(RSType* rs) const
-{
-  if (rs->Naive())
-    rs->Train(std::move(referenceSet));
-  else
   {
-    std::vector<size_t> oldFromNewReferences;
-    typename RSType::Tree* tree =
-        new typename RSType::Tree(std::move(referenceSet), oldFromNewReferences,
-        leafSize);
-    rs->Train(tree);
-
-    // Give the model ownership of the tree and the mappings.
-    rs->treeOwner = true;
-    rs->oldFromNewReferences = std::move(oldFromNewReferences);
+    rs.Search(std::move(querySet), range, neighbors, distances);
   }
 }
 
@@ -373,8 +129,117 @@ void RSModel::serialize(Archive& ar, const uint32_t /* version */)
   if (cereal::is_loading<Archive>())
     delete rSearch;
 
-  // We'll only need to serialize one of the model objects, based on the type.
-  ar(CEREAL_VARIANT_POINTER(rSearch));
+  // Avoid polymorphic serialization by explicitly serializing the correct type.
+  switch (treeType)
+  {
+    case KD_TREE:
+      {
+        LeafSizeRSWrapper<tree::KDTree>& typedSearch =
+            dynamic_cast<LeafSizeRSWrapper<tree::KDTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+    case COVER_TREE:
+      {
+        RSWrapper<tree::StandardCoverTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::StandardCoverTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case R_TREE:
+      {
+        RSWrapper<tree::RTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::RTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case R_STAR_TREE:
+      {
+        RSWrapper<tree::RStarTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::RStarTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case BALL_TREE:
+      {
+        LeafSizeRSWrapper<tree::BallTree>& typedSearch =
+            dynamic_cast<LeafSizeRSWrapper<tree::BallTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+    case X_TREE:
+      {
+        RSWrapper<tree::XTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::XTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case HILBERT_R_TREE:
+      {
+        RSWrapper<tree::HilbertRTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::HilbertRTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case R_PLUS_TREE:
+      {
+        RSWrapper<tree::RPlusTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::RPlusTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case R_PLUS_PLUS_TREE:
+      {
+        RSWrapper<tree::RPlusPlusTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::RPlusPlusTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case VP_TREE:
+      {
+        RSWrapper<tree::VPTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::VPTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case RP_TREE:
+      {
+        RSWrapper<tree::RPTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::RPTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+
+    case MAX_RP_TREE:
+      {
+        RSWrapper<tree::MaxRPTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::MaxRPTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+    case UB_TREE:
+      {
+        RSWrapper<tree::UBTree>& typedSearch =
+            dynamic_cast<RSWrapper<tree::UBTree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+    case OCTREE:
+      {
+        LeafSizeRSWrapper<tree::Octree>& typedSearch =
+            dynamic_cast<LeafSizeRSWrapper<tree::Octree>&>(*rSearch);
+        ar(CEREAL_NVP(typedSearch));
+        break;
+      }
+  }
 }
 
 } // namespace range
